updates based on PR reviews
