+++ translations/ja.yml
@@ -22,7 +22,10 @@ src:
           '4035057905': 画像URL
           '4070151372': 選択
           AssetContainer:
+            '1982191895': ''
             '2409515782': まだアセットがアップロードされていません。アップロードボタンをクリックしてファイルをアップロードしてください。
+            '2854191800': ''
+            '2906553848': ''
             '3057159632': 画像アップロード
             '3715622784': まだ画像がアップロードされていません。アップロードボタンをクリックして画像をアップロードしてください。
             '3863945965': ファイルアップロード
@@ -318,11 +321,12 @@ src:
                 '730907263': インストール済み
                 '3480550112': インストール
             PluginSection:
+              '3060204202': ''
               PluginInstall:
-                '152450576': アップロード済みプラグイン
+                '149186996': ''
+                '311550959': ''
+                '1361263844': ''
                 '1889877734': プラグインリスト
-                '1942751261': GitHub private repository URL
-                '2511713296': GitHub private repository URL
                 '3129492077': PCからZipファイルをアップロード
               PluginList:
                 '1933254449': 開発者向けオプション
